requirements.txt
# EvoHAF: Evolutionary Heterogeneity Analysis Framework
# Requirements for training and inference

# Deep Learning Framework
torch>=1.10.0
torchvision>=0.11.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0

# Medical Imaging
nibabel>=3.2.0
SimpleITK>=2.1.0

# Data Processing
pyyaml>=5.4.0
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0

# Progress Bar
tqdm>=4.62.0

# Logging
tensorboard>=2.8.0
